data/investment_banking_tracker.py
import logging
from typing import Dict, Any
from datetime import datetime

class InvestmentBankingTracker:
    """
    Handles real-time tracking of investment banking trades.
    """
    
    def __init__(self, config: Dict[str, Any]):
        """
        Initialize the InvestmentBankingTracker with configuration settings.
        
        Args:
            config: Dictionary containing configuration parameters
        """
        self.config = config
        self.logger = logging.getLogger(__name__)
        self.logger.info("Initializing InvestmentBankingTracker")

    async def fetch_investment_banking_trades(self) -> Dict[str, Any]:
        """
        Fetch real-time investment banking trades data.
        
        Returns:
            Dictionary containing investment banking trades data
        """
        try:
            self.logger.info("Fetching investment banking trades data")
            # Implementation for fetching investment banking trades data would go here
            # This would typically involve API calls to the data provider
            
            # Placeholder return
            return {
                "timestamp": datetime.now().isoformat(),
                "trades": []
            }
            
        except Exception as e:
            self.logger.error(f"Error fetching investment banking trades data: {str(e)}")
            raise

    async def process_investment_banking_trades(self, trades_data: Dict[str, Any]) -> None:
        """
        Process investment banking trades data.
        
        Args:
            trades_data: Dictionary containing investment banking trades data
        """
        try:
            self.logger.info("Processing investment banking trades data")
            # Implementation for processing investment banking trades data would go here
            
        except Exception as e:
            self.logger.error(f"Error processing investment banking trades data: {str(e)}")
            raise
